pyproject.toml
[project]
name = "code-snippets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "agno>=2.2.12",
    "anthropic>=0.72.1",
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.9.1",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.121.2",
    "huggingface-hub>=0.36.0",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.3",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "langgraph-cli>=0.4.7",
    "langsmith>=0.4.42",
    "markdown>=3.10",
    "markitdown[all]>=0.1.3",
    "mcp>=1.21.1",
    "mlx-embedding-models>=0.0.11",
    "mlx-lm>=0.28.3",
    "mlx-whisper>=0.4.3",
    "ollama>=0.6.1",
    "parakeet-mlx>=0.4.0",
    "psycopg[binary]>=3.2.12",
    "pypdf>=6.2.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
    "yt-dlp[default]>=2025.11.12",
]
